--- apps/web/app/(dashboard)/guest/chats/page.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const page = () => {
-  return <div>page</div>;
-};
-
-export default page;
