--- apps/web/app/(dashboard)/profile/page.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import { Profile } from "../../../components/profile/Profile";
-
-export default function ProfilePage() {
-  return (
-    <div>
-      <Profile />
-    </div>
-  );
-}
