--- apps/web/app/(dashboard)/profile/settings/page.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import { Settings } from "../../../../components/profile/settings/Settings";
-
-export default function SettingsPage() {
-  return (
-    <div>
-      <Settings />
-    </div>
-  );
-}
