+++ apps/web/app/[accessType]/events/[eventSlug]/taskProgressBar.tsx
