+++ apps/web/lib/utils.ts
@@ -1,0 +1,6 @@
+export const UserType = {
+  admin: "ADMIN",
+  planner: "PLANNER",
+  vendor: "VENDOR",
+  guest: "GUEST",
+};
